# frontend/public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "DailyTools247 API",
    "description": "Comprehensive set of free utility and AI-powered APIs for developers, including text analysis, security checkers, formatting utilities, social media copy generation, and email marketing analytics.",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "https://www.dailytools247.app/api",
      "description": "Production Server"
    }
  ],
  "paths": {
...
    "/security/password-generator": {
      "post": {
        "summary": "Generate cryptographically secure passwords",
        "description": "Creates random strings with custom lengths and character types, suitable for high-security applications.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "length": { "type": "integer", "default": 16 },
                  "uppercase": { "type": "boolean", "default": true },
                  "lowercase": { "type": "boolean", "default": true },
                  "numbers": { "type": "boolean", "default": true },
                  "symbols": { "type": "boolean", "default": true },
                  "count": { "type": "integer", "default": 1 }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "result": {
                      "type": "object",
                      "properties": {
                        "password": { "type": "string" },
                        "passwords": {
                          "type": "array",
                          "items": { "type": "string" }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/security/ai-password-strength-explainer": {
      "post": {
        "summary": "Analyze and explain password security and vulnerabilities",
        "description": "Performs password entropy analysis and uses AI to provide actionable advice on how to improve its strength.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["password"],
                "properties": {
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "result": {
                      "type": "object",
                      "properties": {
                        "score": { "type": "integer" },
                        "crackTime": { "type": "string" },
                        "feedback": { "type": "string" },
                        "suggestions": { "type": "array", "items": { "type": "string" } }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/security/ai-text-redaction": {
      "post": {
        "summary": "Redact personally identifiable information (PII) from text",
        "description": "Automatically scans text inputs and redacts sensitive data (emails, phone numbers, IP addresses, names) using AI.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["text"],
                "properties": {
                  "text": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "result": {
                      "type": "object",
                      "properties": {
                        "redactedText": { "type": "string" },
                        "entitiesFound": { "type": "array", "items": { "type": "string" } }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/security/ai-qr-phishing-scanner": {
      "post": {
        "summary": "Scan QR codes and evaluate phishing risk",
        "description": "Extracts URLs from uploaded images of QR codes and runs them through AI analysis to detect potential phishing or malware risks.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["qrUrl"],
                "properties": {
                  "qrUrl": { "type": "string", "description": "URL extracted from QR code" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "result": {
                      "type": "object",
                      "properties": {
                        "isSafe": { "type": "boolean" },
                        "riskScore": { "type": "integer" },
                        "malwareDetected": { "type": "boolean" },
                        "phishingDetected": { "type": "boolean" },
                        "explanation": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/security/ai-url-reputation-checker": {
      "post": {
        "summary": "Check URL safety and reputation",
        "description": "Audits a domain or URL against global reputation, spam, and malware databases.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["url"],
                "properties": {
                  "url": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "result": {
                      "type": "object",
                      "properties": {
                        "domain": { "type": "string" },
                        "reputation": { "type": "string" },
                        "score": { "type": "integer" },
                        "categories": { "type": "array", "items": { "type": "string" } }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}